FreePaper/requirements.txt
flask
twilio
python-dotenv
configparser
fbchat